+++ packages/checkbox/src/types.d.ts
@@ -1,0 +1,7 @@
+export interface CheckboxProps extends React.HTMLProps<HTMLInputElement> {
+  checked?: boolean;
+  disabled?: boolean;
+  intermediate?: boolean;
+  reducedTouchTarget?: boolean;
+  value?: boolean;
+}