+++ packages/hanging-garden/src/HangingGarden.tsx
@@ -28,6 +28,7 @@ import { makeStyles, createStyles } from '@equinor/fusion-react-styles';
  * @param headerHeight The height of the column header. Defaults to 32,
  * @param provideController Returns a ref. this contains the renderGarden function. Used to trigger rerenders at will.
  * @param backgroundColor Backgroun color for the garden. Defaults to  white(0xffffff),
+ * @param disableScrollToHighlightedItem Per default garden centers column of clicked item. This disables that interaction.
  */
 
 const useStyles = makeStyles(() =>
@@ -37,6 +38,9 @@ const useStyles = makeStyles(() =>
       flex: '1 1 auto',
       height: '100%',
       minWidth: 0,
+      minHeight: 0,
+      maxWidth: '100%',
+      maxHeight: '100%',
     },
   })
 );
@@ -55,7 +59,9 @@ function HangingGarden<T extends HangingGardenColumnIndex>({
   headerHeight = DEFAULT_HEADER_HEIGHT,
   provideController,
   backgroundColor = 0xffffff,
-}: HangingGardenProps<T>) {
+  colorMode = 'Regular',
+  disableScrollToHighlightedItem = false,
+}: HangingGardenProps<T>): JSX.Element {
   const [maxRowCount, setMaxRowCount] = useState(0);
   const [expandedColumns, setExpandedColumns] = useState<ExpandedColumns>({});
 
@@ -63,7 +69,7 @@ function HangingGarden<T extends HangingGardenColumnIndex>({
   const canvas = useRef<HTMLCanvasElement>(null);
   const stage = useRef<PIXI.Container>(new PIXI.Container());
 
-  const scroll = useScrolling<T>(canvas, container, itemKeyProp);
+  const scroll = useScrolling<T>(canvas, container, itemKeyProp, disableScrollToHighlightedItem);
   const textureCaches = useTextureCaches();
   const popover = usePopover();
 
@@ -103,6 +109,7 @@ function HangingGarden<T extends HangingGardenColumnIndex>({
             renderItemContext,
             renderHeaderContext,
             popover,
+            colorMode,
           }}
         >
           <Garden<T> provideController={provideController} />