+++ packages/hanging-garden/src/models/HangingGarden.ts
@@ -12,6 +12,8 @@ export type GardenController = {
   clearGarden: () => void;
 };
 
+export type ColorMode = 'Regular' | 'Color blind';
+
 export type HangingGardenProps<T extends HangingGardenColumnIndex> = {
   columns: HangingGardenColumn<T>[];
   highlightedColumnKey: string | null;
@@ -26,4 +28,6 @@ export type HangingGardenProps<T extends HangingGardenColumnIndex> = {
   renderHeaderContext: (key: string, context: HeaderRenderContext) => void;
   provideController?: MutableRefObject<GardenController | null>;
   backgroundColor?: number;
+  colorMode?: ColorMode;
+  disableScrollToHighlightedItem?: boolean;
 };