+++ packages/hanging-garden/src/renderHooks/useScrolling.tsx
@@ -23,7 +23,8 @@ export type Scroll<T extends HangingGardenColumnIndex> = {
 const useScrolling = <T extends HangingGardenColumnIndex>(
   canvas: RefObject<HTMLCanvasElement> | null,
   container: RefObject<HTMLDivElement> | null,
-  itemKeyProp: keyof T
+  itemKeyProp: keyof T,
+  disableScrollToHighlightedItem?: boolean
 ): Scroll<T> => {
   const isScrolling = useRef(false);
   const scrollTop = useRef(0);
@@ -54,11 +55,16 @@ const useScrolling = <T extends HangingGardenColumnIndex>(
         return false;
       }
 
-      scrollLeft.current =
+      //Calculate how far from the left to scroll, while ensuring scroll is not below 0.
+      const scrollWindowTo = Math.max(
         highlightedColumnIndex >= 0
           ? (container.current.scrollLeft =
               highlightedColumnIndex * itemWidth - container.current.offsetWidth / 2 + itemWidth / 2)
-          : 0;
+          : 0,
+        0
+      );
+
+      scrollLeft.current = container.current.scrollLeft = scrollWindowTo;
 
       if (canvas?.current)
         canvas.current.style.transform = `translate(${scrollLeft.current}px, ${scrollTop.current}px)`;
@@ -79,7 +85,7 @@ const useScrolling = <T extends HangingGardenColumnIndex>(
 
   const scrollToHighlightedItem = useCallback(
     (columns: HangingGardenColumn<T>[], highlightedItem: T | null, itemWidth: number): boolean => {
-      if (!highlightedItem) return false;
+      if (disableScrollToHighlightedItem || !highlightedItem) return false;
       const highlightedIndex = columns.findIndex((column) =>
         column.data.some((item) => {
           return item[itemKeyProp] === highlightedItem[itemKeyProp];
@@ -88,7 +94,7 @@ const useScrolling = <T extends HangingGardenColumnIndex>(
 
       return scrollTo(highlightedIndex, itemWidth);
     },
-    [scrollTo, itemKeyProp]
+    [scrollTo, itemKeyProp, disableScrollToHighlightedItem]
   );
 
   return {