+++ packages/radio/src/Radio.tsx
@@ -1,0 +1,9 @@
+import { FunctionComponent } from 'react';
+import { RadioProps } from './types';
+import '@equinor/fusion-wc-radio';
+
+export const Radio: FunctionComponent<RadioProps> = (props) => {
+  return <fwc-radio {...props}></fwc-radio>;
+};
+
+export default Radio;