+++ packages/styles/src/index.ts
@@ -1,5 +1,6 @@
 export * from './make-styles';
 export * from './theme';
 
-export { withStyles, createStyles } from '@material-ui/styles';
+export { withStyles, createStyles, Styles } from '@material-ui/styles';
 export { default as clsx } from 'clsx';
+export { ThemeProvider } from './ThemeProvider';