+++ packages/hanging-garden/CHANGELOG.md
@@ -3,6 +3,89 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.4.2 (2021-07-05)
+
+**Note:** Version bump only for package @equinor/fusion-react-hanging-garden
+
+
+
+
+
+## [1.4.1](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-hanging-garden@1.4.0...@equinor/fusion-react-hanging-garden@1.4.1) (2021-06-30)
+
+
+### Bug Fixes
+
+* ensure that the scroll to calcualtion does not endup being below 0 ([1f64d86](https://github.com/equinor/fusion-react-components/commit/1f64d861d00cb7eb9d349bdbc1ef78c0b49ac9aa))
+
+
+
+
+
+# [1.4.0](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-hanging-garden@1.3.0...@equinor/fusion-react-hanging-garden@1.4.0) (2021-06-28)
+
+
+### Bug Fixes
+
+* a minor hack to get control over the caches ([20a14df](https://github.com/equinor/fusion-react-components/commit/20a14df2391f21487eafcb63bd5bc63bc778db62))
+* handle onload error ([3813f5e](https://github.com/equinor/fusion-react-components/commit/3813f5e04e7c728bbe9a9c23557106b60d4f1b67))
+
+
+### Features
+
+* added a new setting: disableScrollToHighlightedItem ([e9bcb0b](https://github.com/equinor/fusion-react-components/commit/e9bcb0b728133ed7b4f82c0c5a3f9d093e32d98c))
+
+
+
+
+
+# [1.3.0](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-hanging-garden@1.2.0...@equinor/fusion-react-hanging-garden@1.3.0) (2021-06-28)
+
+
+### Bug Fixes
+
+* a minor hack to get control over the caches ([20a14df](https://github.com/equinor/fusion-react-components/commit/20a14df2391f21487eafcb63bd5bc63bc778db62))
+* handle onload error ([3813f5e](https://github.com/equinor/fusion-react-components/commit/3813f5e04e7c728bbe9a9c23557106b60d4f1b67))
+
+
+### Features
+
+* added a new setting: disableScrollToHighlightedItem ([e9bcb0b](https://github.com/equinor/fusion-react-components/commit/e9bcb0b728133ed7b4f82c0c5a3f9d093e32d98c))
+
+
+
+
+
+# 1.2.0 (2021-06-28)
+
+
+### Features
+
+* add color blind mode to hanging carden component and cache key ([53aea21](https://github.com/equinor/fusion-react-components/commit/53aea2115f2d8907e61661a8bb256fe4c519c6b9))
+
+
+
+
+
+## [1.1.3](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-hanging-garden@1.1.2...@equinor/fusion-react-hanging-garden@1.1.3) (2021-06-23)
+
+**Note:** Version bump only for package @equinor/fusion-react-hanging-garden
+
+
+
+
+
+## 1.1.2 (2021-06-17)
+
+
+### Bug Fixes
+
+* fixed linting ([aa160f8](https://github.com/equinor/fusion-react-components/commit/aa160f8a460256fe7cc86947d031826b34a190c4))
+
+
+
+
+
 ## [1.1.1](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-hanging-garden@1.1.0...@equinor/fusion-react-hanging-garden@1.1.1) (2021-03-23)
 
 **Note:** Version bump only for package @equinor/fusion-react-hanging-garden