+++ packages/styles/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.2.0 (2021-07-05)
+
+
+### Features
+
+* added theme provider ([a7e3e3f](https://github.com/equinor/fusion-react-components/commit/a7e3e3f0657ce4d5659bb387ca71d823242df20f))
+
+
+
+
+
+## 0.1.3 (2021-06-28)
+
+**Note:** Version bump only for package @equinor/fusion-react-styles
+
+
+
+
+
+## [0.1.2](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-styles@0.1.1...@equinor/fusion-react-styles@0.1.2) (2021-06-23)
+
+
+### Bug Fixes
+
+* fixed css styling issue ([809e1bf](https://github.com/equinor/fusion-react-components/commit/809e1bfaac0a99d0bda3b32d49d51e2043428171))
+
+
+
+
+
+## 0.1.1 (2021-06-17)
+
+
+### Bug Fixes
+
+* fixed linting ([0278fb0](https://github.com/equinor/fusion-react-components/commit/0278fb00da66f4cf6be855fc66eaa06074806465))
+
+
+
+
+
 # [0.1.0](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-styles@0.1.0-alpha.2...@equinor/fusion-react-styles@0.1.0) (2021-03-23)
 
 **Note:** Version bump only for package @equinor/fusion-react-styles